apps/cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gridone-devices-manager",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
]

[project.scripts]
gridone = "cli.app:app"

[tool.uv]
workspaces=true

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "cli"
module-root = "src"


[tool.pytest.ini_options]
pythonpath = ["src"]
